Move tmpdir test helper to test helpers file
